refactor: rename tasklist to list
A custom component should be multi-word so as not to conflict with standard html components. In order to achieve this before we created "TaskList" components. This is a bit of a misnomer because it doesn't match the original entity modeling and additionally, our Vuex store state was mapping tasks when they should be called lists. This refactoring puts us back to where we should be. We're just prefixing our components with our product acronym which is best practice anyways.